requirements.txt
annotated-types==0.6.0
anyio==4.1.0
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
Flask==3.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
replicate==0.21.1
sniffio==1.3.0
typing_extensions==4.8.0
Werkzeug==3.0.1
